requirements.txt
# Dependencias para la aplicación web con Streamlit

streamlit>=1.0.0
pandas>=1.3.0
openpyxl>=3.0.0
python-barcode>=0.13.0
Pillow>=8.0.0
pywin32>=300
